feat: add support for robots.txt generation
This commit adds functionality for querying the robots.txt
API endpoint and saving the generated robots.txt content.
A new block `robots_txt` is added for granular configuration
of generation options. If the block is present, a single
query is performed during the provisioning phase of the module
and cached for subsequent use. In a HTTP request context, the
variable `http.vars.dv_robots_txt` is avaliable to access the
cached robots.txt content.

The README.md is updated with documentation for the syntax of the
HTTP handler directive and notes about supported features.